indexing/requirements.txt
pandas
python-dotenv
sentence-transformers
requests
PyYAML
opensearch-py
confluent-kafka